Move components up the tree
