+++ marketplace/cloudinary-assets/src/AppConfig/AppConfig.spec.tsx
